Refactoring of the agent command's command classes
