+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/command/CommandDelegator.java
@@ -3,7 +3,7 @@ package rocks.inspectit.ocelot.core.command;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import rocks.inspectit.ocelot.commons.models.command.Command;
-import rocks.inspectit.ocelot.commons.models.command.response.CommandResponse;
+import rocks.inspectit.ocelot.commons.models.command.CommandResponse;
 import rocks.inspectit.ocelot.core.command.handler.CommandExecutor;
 
 import java.util.List;
